requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pgvector
sentence-transformers
numpy
pydantic
python-dotenv
